Change the validation strategy (#5)
* [Feature]: revise the training pipeline

* [Fix]: Rework validation strategy through a regular split

* [Feature]: Save metrics to an excel file

* [Fix]: exclude nullified values from MAPE calculation

* [Fix]: change the original dataset

* [Fix]: add metrics for multiple trainings

* [Bugfix]: exclude NaNs from the train/val indices

* [Fix]: change data type

+++ inference.py
@@ -23,7 +23,7 @@ from tools.model import Regressor
 
 
 def main(
-    data: Union[List[int], np.ndarray, pd.DataFrame, str],
+    data: Union[List[float], np.ndarray, pd.DataFrame, str],
     lumen_model_path: str,
     stress_model_path: str,
     use_golden_features: bool,
